refactor: Move DefaultMVConfig and MVConfigNodes to config package
